feat: 生成中使用tqdm可视化进度

+++ llm_inference/translation/seed_x.py
@@ -23,7 +23,7 @@ def translate(sentence: str, target_lang: str = 'en', resample: int = 1, **kwarg
               f'<sentence>{sentence}</sentence>'
               f'<{target_lang.lower().replace("<", "").replace(">", "")}><sentence>')
     logger.debug(f'REPROMPT WITH BEST ANS: {prompt.replace("\n", " ")}')
-    kwargs['max_tokens'] = len(sentence) * 2.5
+    kwargs['max_tokens'] = int(len(sentence) * 2.5)
     best_ans = sorted(inference(prompt=[prompt] * resample, llm=seed_x_lm, regex=regex, **kwargs),
                       key=lambda x: len(x[0]), reverse=True)[0].strip()
     best_ans = best_ans[:best_ans.rfind('</sentence>')].rstrip()
